requirements.txt
# Core Framework
streamlit>=1.28.0
requests>=2.31.0

# Chemistry & Molecular
rdkit>=2023.9.1

# Bioinformatics
biopython>=1.81

# AI/ML
google-generativeai>=0.3.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
